feat(v0.2.0): Config construction, Print API, example sugar pass
Ordinary CLIs use evo.New() / New(Config) with TTY and NO_COLOR defaults,
fmt-style Print/Printf/Println and Verbose, Task.Done(summary), Progress(int),
and Itemf/Taskf. Option-based construction is NewWithOptions. Examples and
docs teach the Print → Item/Task → Capture adoption ladder.

+++ examples/live-progress/main.go
@@ -128,7 +128,7 @@ func main() {
 
 	entries, _ := os.ReadDir(dir)
 	for i := range entries {
-		scan.Progress(int64(i+1), int64(len(entries)))
+		scan.Progress(i+1, len(entries))
 		done := int64(float64(totalBytes) * float64(i+1) / float64(len(entries)))
 		download.Bytes(done, totalBytes)
 		time.Sleep(stepDur)
